Bugfix for issue #37
This is a fix of issue #37. 

This issue is caused by an error of SDP. When the current tree type is luma, set_blk_skip() will set chroma planes to an uninitialized state for further checking in debug mode, but it will not do this operation in release mode. Since av1_encode_sb() does not take into account the current tree type, the uninitialized blk_skip of chroma planes may be used even if the current tree type is luma.