Bugfix with joint mvd on and smvp improve off
array cm->ref_frame_relative_dist is defined under compiler flag SMVP_IMPROVEMENT, but extensively used for joint mvd coding. So when SMVP_IMPROVEMENT is off but joint mvd is on, there is building errors. To solve this problem, this variable is defined either smvp improvement or joint mvd on.