B091 New Reference Signaling
This experiment removes the named references and replaces them by ranked references. Particularly, rankings of all available reference frames can be derived by encoder and decoder normatively. Only the ranks of the reference frames will be signaled, which simplifies the bitstream syntax.

The purpose of experiment is to introduce a more flexible framework, which opens the door to new coding tools and optimization algorithms. For example, it is now easily to increase the number of available reference frames in this new framework.

STATS_CHANGED