revert sign initialization for unit test
This commit reverses the sign zero initialization for C function for a nightly-test issue reported additionally in #73, it reverts a part of the previous MR !306 (merged) (which originally passed the unit-tests). The reverted change does not impact coding gain and is the simplest fix. Moreover, there are no valgrind errors since SSE/AVX fixes are kept in.