Remove disabled low bit-depth codepath
This MR attempts to remove the low bitdepth codepath and the public API related to it.

AOM_IMG_FMT_HIGHBITDEPTH is left in place in order to let the library users feed the encoder frames using 8bit buffers for 8bit content. These are in turn upshifted to 16bit buffers in the aom_codec_encode_fn_t callback from aom_codec_av1_cx_algo to be handled internally (the aomenc CLI is already doing the same, but other library users do not benefit from it).