Make av1_convert_qindex_to_q() return real u-qstep / Fix use-fixed-qp-offsets=2
1. This patch makes av1_convert_qindex_to_q() return unitary
normalized quantization step size to avoid confusion in the
future. Previously the function was returning 2 times the
normalized stepsize. This function is used mostly in rate control
code that is not exercised in the CTC, but this change will
facilitate future work on rate/quality control improvements,
by removing hidden scale factors.

STATS_CHANGED may happen for some non-CTC tests.

2. Fix --use-fixed-qp-offsets=2 option