Turn on the compiler flag for CWG-C063
In the previous MR !481, both the compiler flags for CWG-C063 and CWG-C061 are turned off for test purpose. And the nightly test run is green and there are no OOM issues for gitlab runners after turning off these two compiler flags. This MR is to turn on the compiler flag for CWG-C063 to further narrow down the problem.

STATS_CHANGED