Check for valid chroma bsize on frame boundary
Currently, the boundary condition does not explicitly check for
the chroma bsize. This can lead to the encoder getting stuck in an
infinite loop when chroma 422 subsampling is used.

This commit adds an explicit check for chroma block size in
`is_partition_implied_at_boundary`. Furthermore, for LUMA_TREE above sdp
threshold, a check is added to ensure that the corresponding chroma
bsize is valid.

See also: #131