ERP: Fix a misidentification of counterpart block
Current, the encoder searches for a counterpart pc_tree node by checking
whether the bsize matches. However, this is insufficient as the
bsize_base might not match, which leads to mismatch in chroma coding.

Below is an example:

```
   +--------+     +---------+
   |        |     |         |
 4 |        |   4 |         |
   +--------+     +----+----+
   |        |     | 4  |    |
 4 |        |     |    |    |
   +--------+   8 |    |    |
   |        |     |    |    |
 8 |        |     +----+----+
   |        |   4 |         |
   |        |     |         |
   +--------+     +---------+
       8               8
```

Here, the first blocks both share a dimension of 8X4. But the first one
is a chroma reference, so its bsize_base is also 8X4. But the second one
is not a chroma reference, so its bsize_base is 8X16.

This commit adds an additional check by comparing the chroma_ref_infos
in addition to bsize.

This fixes a couple mismatches and a tsan failure. See #132.

STATS_CHANGED