CCTX: Disable CCTX when TREE_TYPE is LUMA_PART
Currently, `is_cctx_allowed` is called by various functions even when
the partition tree type is `LUMA_PART` (e.g. by `av1_xform_quant`). When
this happens, the `chroma_plane_bsize` variable could be `BLOCK_INVALID`,
which leads to out-of-found indexing later on.

This commit add a check to so that `is_cctx_allowed` returns 0
immediately if the current tree_type is `LUMA_PART`.

No stats difference is observed by this change.

See #132.