CWG-D027 (compound weighted prediction)
This contribution proposes an extension to the COMPOUND_AVERAGE mode that enables adaptive weighting of the two prediction signals in compound prediction mode. Specifically, when the two reference frames are from different directions, five weighting factors {8, 10, 6, 12, 4} are supported. In contrast, when the reference frames are from the same direction, five different weighting factors {8, 12, 4, 20, -4} are supported.

STATS_CHANGED