Fix issue #156
Fix the nightly test failures after merging CWG-D027 (compound weighted prediction).
The reason for the nightly test failure is that function of calculating RD cost is accessing the frame level structure, which is needed. Actually, there is no need to access the frame level structure in this function, so this code is removed in the commit.

No performance changes to the bitstream.