Fix alignment on insp_sb_data.
Coeffs do not need to be 32-byte aligned for inspection purposes.

Coeff alignment is needed within the codec for SIMD optimizations, but this isn't relevant for inspection, which just has a copy of the coeffs and performs no calculations on them.

The allocation of the inspection superblock grid (`insp_sb_data *sb_grid`) was not taking this alignment into account, causing elements past the first one to be unaligned, which resulted in a UBSAN error. Removing the alignment on `insp_sb_data` fixes this.