Disable global motion by default
At the CWG meeting on June 30th 2023, we decided to disable global
motion in the default configuration for the moment. This is because
currently it does not meet the required speed:quality tradeoff,
but it may be re-enabled in future if this is improved sufficently.

Impact of global motion on vs. off (AV2 CTC, RA config):

research-v4.0.0: +0.055% BDRATE, +1% encode time

research-v4.0.0 + CWG-C034 Improved Global Motion (candidate, but
not yet merged): -0.001% BDRATE, +2% encode time

STATS_CHANGED