CWG-D096: Explicit signaling of a secondary transform set
This MR has two parts:
1) A prep work, Extend the data type of TX_TYPE from uint8_t to uint16_t
2) Actual work for D096, "Explicit signaling of a secondary transform set"
-------------------------------------------------------------------------
Actual work for D096:
Explicit signaling of a secondary transform set (D096)

The Intra Secondary Transform (IST) coding tool has been
merged into AVM and enabled by default. It implicitly
derives the secondary transform set based on the context
of intra prediction directions and explicitly signals
the kernel type within a set.

This patch enables signaling the secondary transform set 
explicitly, such that any secondary transform set can be
selected for a block regardless of the intra prediction
mode being applied.

If the secondary tx is disabled for the block, i.e. kernel type
is zero, then the set is not signaled.

By default, the CONFIG_IST_SET_FLAG flag
is turned on. 

STATS_CHANGED

-------------------------------------------------------------------------
A prep work:
Extend the data type of TX_TYPE from uint8_t to uint16_t

This is a prep work for the explicit signaling of
secondary transform set (IST).

The TX_TYPE has been used for internal representation of the
primary and secondary transform kernel type.

Then, the soon coming change in separate commit is to
explicitly signal the set of secondary transform.

This is a lossless change in terms of coding efficiency,
and should minimally affect on both encoding and decoding time.