Fix issue #223
In function is_smooth(), when SDP is on, luma and chroma block may have different intraBC flag, so it is intended to check the corresponding intraBC flag for each plane rather than always checking the intraBC flag of luma block.

This incorrect behavior is corrected in this commit.

STATS_CHANGED.