CWG-D048 (Explicit BAWP)
In this MR, when one block is coded as BAWP mode, another flag is signaled to indicate whether explicit or implicit signaling for BAWP scaling factors is used. When explicit signaling of BAWP scaling factors is employed, another flag is signaled to indicate which scaling factor is used for current block and the offset is derived based on linear equation.

Compile time flag CONFIG_EXPLICIT_BAWP is turned off by default.