Fix av1_scale_warp_model if in_distance == 0
Per the linked bug report, it is possible for in_distance == 0 to occur in a
valid bitstream. But the code currently fails with an assertion failure in
this case.

Replace the assertion by properly handling in_distance == 0. The only reasonable
behaviour is to use the identity model as a reference. In addition, the input
model should always equal the identity model in this case, so we add checks for
this condition.

Fixes https://gitlab.com/AOMediaCodec/avm/-/issues/256