Fix for short_refresh_frame_flags = 0 case and use bit masks to represent refresh frame flags
(1) We need to read refresh frame flags when short_refresh_frame_flags = 0.

Fixes issue #292.

(2) In AV1, refresh frame flags are represented by bit masks. In the original CWG-D044, refresh frame flags are represented by index.

In this change, we modify CWG-D044 such that refre frame flags are also represented by bit masks to make it consistent with AV1.

The savings introduced by CWG-D044 remain unchanged.

More explanation:

There are 4 different cases of refresh frame flags. (1). Refresh all frames in the buffer. (2). Do not refresh any frame. (3). Refresh one frame in the buffer. (4). Refresh more than one frame in the buffer.

In AV1, a bit mask is used to represent refresh frame flags. If the i-th frame is to be refreshed, the i-th bit is set to 1. Other bits are zero. For example, for case (1), bit mask is 0xFF. For case (2), bit mask is 0. For case (3), bit mask of refreshing the 6th frame is 0b00100000. For case (4), bit mask of refreshing the 1st and 2nd frame is 0b00000011.

In CWG-D044, if enable_short_refresh_frame_flags is true, it signals 3 bits instead of 8 bits to indicate refresh frame flag, and it can only support case (1), (2), (3). By design, if the encoder wants to refresh more than one frame (case 4), enable_short_refresh_frame_flags has to be false.

In the original implemetation of CWG-D044, a index (0-7) is used to indicate frames to be refreshed. For example, if the 6th frame is to be refreshed, we signal 5 (start from 0).

Although this works well, it creates extra complexity for the code. When enable_short_refresh_frame_flags is true, the refresh frame flag is represented by an index. When enable_short_refresh_frame_flags is false, the refresh frame flag is represented by a bit mask.

This change unifies them. Keep the design the same as AV1 by also using a bit mask to represent the refresh frame flag.