CWG-D165: Disable OBMC by default
Set default value of --enable-obmc argument to 0. This disables OBMC in the default
configuration, but still allows it to be manually enabled for future testing.

Per discussion in the Codec Working Group, add a test which explicitly enables OBMC
to ensure that the encoder is not broken by later changes. It turns out that
CpuSpeedTest (cpu_speed_test.cc) already does everything what we want, including
testing special configurations like lossless and screen content. So we can simply
duplicate that test with OBMC enabled.

STATS_CHANGED