Fix BAWP code when pred area falls beyond valid ref frame boundary
This change is to fix #333

In parent version, weighted prediction was applied only to the part
of the block that is within valid frame area and remaining pred block
from padded region is not compensated. In this CL, weighted prediction
is done for the entire block in this case.

The average BD-Rate(avg psnr) impact is -0.0028% over RA test-set for
33/65 frames.