REFINEMV: fix the right/bottom coordinate of the padded ref area
This CL is from Deepa, and it fixes #353

In the function 'refinemv_highbd_pad_mc_border()', the pixel
to be padded on the right border of the block is obtained from
'ref_row[ref_area->pad_block.x1 - 1]'.  When 'ref_area->pad_block.x1'
is 0, the pixel is fetched from 'ref_row[- 1]'.  In the decoder,
as the reference frame buffers are not padded, 'ref_row[- 1]' is
uninitialized.

As 'ref_area->pad_block.x1/y1' is one greater than the right/bottom
coordinate of the padded reference area, the lower bound
for 'ref_area->pad_block.x1/y1' is changed to 1.