CWG-E014: MVD coding
In this commit, instead of separate coding of the row and column components of a MVD, a joint vector coding method is proposed. In the proposed method, the magnitude of the MVD is coded by two parts, sum of absolute values of row and column MVD ( called shell index) and
column MVD. The proposed MVD coding replaced the existing MVD coding method including for IBC mode.

In addition, instead of the explicit signaling , under certain conditions, the sign of the last non-zero MVD component is derived from the parity of the magnitude of the decoded MVDs of the same block.

STATS_CHANGED