Don't enable CONFIG_DEBUG for RelWithDebInfo
This is a port of the libaom CL
https://aomedia-review.googlesource.com/c/aom/+/122801.

Don't enable CONFIG_DEBUG if CMAKE_BUILD_TYPE is RelWithDebInfo. Only
enable CONFIG_DEBUG if CMAKE_BUILD_TYPE is Debug.

The code enabled by CONFIG_DEBUG is debugging code and therefore it may
change the behavior of the binary. The goal of the RelWithDebInfo build
type is to add debugger symbol information to a release build, so it
should not enable debugging code. (For example, RelWithDebInfo disables
assertions by -DNDEBUG.)