Encoder optimization: Refactor CCTX angle evaluation code
In parent version, RD evaluation of CCTX angles was partially skipped
based on eob of c1 component and the relevant conditions were
evaluated only after both c1 and c2 are processed. In this CL, the
processing of c2 is also conditionally skipped based on eob through
refactoring and the relevant conditions are cleaned-up appropriately.
This CL produces bit-exact output with instruction count reduction of
0.2%-0.3%.