CWG-E006: Linear intra prediction with matched reference templates
We introduce a new mode for intra prediction. Only when intra block
copy is used, this mode can be selected.

It derives a linear equation from matched reference templates, and
applies this linear equation to get the prediction of the current
block.

Signal a binary flag mbmi->morph_pred (0 or 1).

Some restrictions applied to the new mode:

Do not allow the proposed mode for block larger than 64

Do not allow the linear model for local intrabc

In this patch, we also unify the linear parameter derivation.

Let the proposed method, implicit cfl and bawp all share the same
function for deriving linear parameters.

Note that the shift used in bawp and the proposed method are the
same, the shift in cfl is different.

Also, bawp has made a sligh modification to numerator and
denominator before division, we keep that modification in this patch.

Change-Id: Id71fd7f2bbd41757d082f418f545f8d9c4aefd3e