Enable E055 + E039 combination
E039 (wide intra angle) was not working on top of E055 (txfm partition
improvement) as is, because E039 doesn't support the case when txfm
partition sizes inside one coding block are different.

In this commit, we add the support to keep track of the wide intra
angle for each txfm partition block, such that the two tools can work
well together.

STATS_CHANGED