Further fix the encoder bug of wide angles
Before fix !1118, after looping all the transform block partitioning, the best TX size is stored, but the best is_wide_angle variable is stored well. For example, if one coded block is square block 16x16, the is_wide_angle variable should be always zero. But after horizontal or vertical transform block partitioning, the value of this variable may be set to 1. But if the best transform block partitioning is still PARTITION_NONE, then the value of is_wide_angle variable needs to be reset to 0. But in !1118, only variable is_wide_angle is reset, but mapped_wide_angle is not set. In this MR, mapped_wide_angle is also reset for the best TX size.