Bug fix of E062 for CfL mode
STATS_CHANGED

This MR fix solves a hidden bug in AVM software for CfL (both explicit and implicit CfL) prediction modes. In current AVM, the luma block size for a CfL block is capped to maximum chroma transform size. This bug is hidden when maximum CfL size is 32x32 in luma block size. The bug is activated in E062 when CFL is enabled up to 64x64 Luma block size.

This fix will further increase the gain of CWG-E062.