Fix energy calculation during CCTX angle RD evaluation
This issue was reported in #375.

CCTX angle evaluation is conditionally skipped based on energy of dequantized coefficients and the dequantized coefficients are stored with 32-bit precision. In parent version, the energy was calculated as if dequantized coefficients are stored in 16-bit buffers. This patch corrects the energy calculation and removes one stale comment as well.

This fixed brought minor performance change: BD rate: +0.012% with 0.1% encoder time increase.

STATS_CHANGED