Move allocation of pred buffer in upsampled_pref_error()
This patch moves the allocation of pred buffer in 'upsampled_pref_error()' function from stack memory to MACROBLOCK structure which improves the function performance for GCC compiler due to better stack memory management.

This is a bit-exact change, and gives 1% - 3% encoder time reduction for GCC compiler (higher speedup is seen for low QPs).

Performance impact on the Clang compiler is negligible.