Fix a thread sanitizer issue
This is the fix for #399.

A recent change in sub-pel motion search allocated the pred buffer in heap memory instead of using the stack memory. In multi-thread case, except the encoder stage, motion search is also done in TPL and first pass encoding. Here, the memory pointers are assigned accordingly in those 2 cases as well.