Optimize dynamic memory allocation in av1_optimize_txb_new()
In the parent version, the coef_info structure was dynamically allocated in the av1_optimize_txb_new() function. As the function is called a large number of times, the heap memory allocation adds a significant overhead which reduces the performance of the function.

In this patch, the structure is moved to macroblock structure such that the overhead due to repetitive memory allocation is avoided.

This is a bit-exact change, and gives an encoder time reduction of 0.75% (AVG) and 1% for Qp 110 & 135.