Update ref frame distances for explicit-ref-frame-map=1
This is a decoder fix when explicit-ref-frame-map=1. After
decoder receives the ref frame indexes, it must update the
ref_frame_distance in cm->ref_frames_info. Otherwise decoder
may use incorrect ref frames to decode.

This fix doesn't change the bitstream in CTC since explicit-ref-frame-map is set to 0 in CTC.