ResizeInternalTestLarge: Correct the max quantizer.
For the test `TestInternalResizeChangeConfig`, the variable
`change_config_` is set to True, so it switches max quantizer during
encoding.

Originally, the max quantizer values were based on the old qindex range
of 0 to 65, where max qindex was being switched from 48 to 60.

For AV2, we later updated the QP range to 0 to 255 (in this MR:
https://aomedia-review.googlesource.com/c/aom/+/114405). However, only
the value 48 was mapped to equivalent value 192, and the mapping of
value 60 to 240 was missed.

We fix that here, with the side benefit of also speeding up this unit
test.