Fix 2D nonsep convolve tests
The random inputs generated were already extended by 4 pixel on all sides. The test however was centering the input by an additional 3 pixels for luma and 2 pixels for chroma without reducing the width and height. This causes invalid pixels to be accessed beyond the right and bottom edges for the luma case. This patch resolves the issue by removing the unnecessary centering.