AVxEncoderParmsGetToDecoder: test on lower resolution
The purpose of the test is to make sure certain high level parameters
given to the encoder are respected by testing those parameters after
encoding and decoding.  So, we can use a lower resolution input video
for the same purpose to speed up the unit test.

Also, use fixed QP encoding.