Set QP explicitly for CQ mode in a number of unit tests.
CQ mode is a mix of fixed QP and VBR modes, so it needs to set both;
otherwise some default values are used, which result in very low QPs.

QP for CQ mode is now explicitly set in following tests:
- test/subgop_test.cc
- test/altref_test.cc
- test/error_resilience_test.cc
- test/fwd_kf_test.cc
- test/kf_test.cc

Also, for subgop_test, set target bitrate for all rate control modes
(instead of using default of 256), so VBR and other modes will use this
explicit setting.

This is to speed-up these unit tests.