OBMCEncodeTest: encode fewer frames for two high bitrate tests.
This is to speed-up these slow unit tests.

Note that the purpose of the tests is to ensure OBMC works correctly for
non-multiple-of-64 wide clip. So, encoding fewer frames will still
ensure this coverage.

Also, the lower bitrate test still encodes same number of frames as
before to ensure more coverage.