Make the interp filters used during motion estimation consistent
In case of compound blocks with only one NEWMV component or
joint MVD coding mode, during motion estimation process, the
interpolation filters used to generate the prediction block for
the fixed MV was set to `mbmi->interp_fltr` (interpolation filter
of the previous search configuration in `handle_inter_mode()`
function). However, the interpolation filter during sub-pel
search is set to `EIGHTTAP_REGULAR` in the function
`av1_get_filter()`. In this patch, the  interpolation filter
used to generate the prediction block for the fixed MV is set to
`EIGHTTAP_REGULAR` making it consistent with the interpolation
filter used during sub-pel search.

STATS_CHANGED

Fixes #441