Handle superres properly with LR frame filters
When superres is used, the tskip values from the coded
resolution will be misaligned with LR for use with classified
wiener filters in PC_WIENER or frame level filters.
In these cases assume tskip values to be 0 without reading
from the tskip buffer.

Resolves Issue #457 related to frame level classified filters.

Besides PC_WIENER is now also enabled when superres is used.
This is a normative change in behavior, but does not affect the CTC.