Fix few encoder bugs of super-res mode
This commit fix few encoder bugs of super-res mode. All of the changes are non-normative. This commit does not change the results of CTC configurations. 
Following bugs are fixed.

1. Fixed the various variance computation functions when scaled reference is enabled.
2. Properly re-allocate superblock size dependent buffers
3. Fix frame level SSE computations 
4. Properly saved and restored necessary data after and before encoding of each resolution
5. Properly restore the scaled prediction buffers and scaling factors during single and compound motion search
6. Adjust QP of the frame based on resolution when super-res is enabled. 

Before the above fixes, there was an encoder crash with super-res mode on. Item (2) of the above list can fix encoder crash but produces large BD rate loss. With all of the fixes above gain of superres mode is neutral.      

STATS_CHANGED