Fix when TIP block is not aligned to 8x8 grid
This change fixes the offsets used during
prediction data generation of TIP_FRAME_AS_REF
blocks which are not aligned to 8x8 grid.

For 420 chroma subsampling, as TIP_FRAME_AS_REF
blocks are always aligned to an 8x8 grid, this
change should be bit-exact.

STATS CHANGED for 422, 444 chroma subsampling

Fixes #477.