Fix ptree initialization
This changes the default partition tree type from `PARTITION_NONE` to `PARTITION_INVALID`, because it caused some issues for blocks out of frame.

I have for example a bitstream of size 96x112, with superblock size = 64 and SDP enabled.

![image](/uploads/8ed7bbf4dd301fce56c95514672acbef/image.png)

In luma part, in the bottom-right superblock the first partition type is `PARTITION_VERT`, and even if the right part is totally out of frame, it gets added to the ptree [here](https://gitlab.com/AOMediaCodec/avm/-/blob/main/av1/decoder/decodeframe.c?ref_type=heads#L2434). That's where the problem begins.

![image](/uploads/4b3ed31dbe8d447862de9f6771d7b3f5/image.png){width="715" height="430"}

Later on (in chroma part), we pass in [tree_has_bsize_smaller_than](https://gitlab.com/AOMediaCodec/avm/-/blob/main/av1/common/av1_common_int.h?ref_type=heads#L2949) function that checks if all block sizes in the luma partition tree are smaller that a given threshold. Notice the first condition that exits early only when the child is null or has a `PARTITION_INVALID`, but not for our block completely out of frame. Since this ptree was calloc'ed, the ptree-\>bsize is equal to 0 (`BLOCK_4x4`), which leads the function to return `true` instead of `false`. I think this is an issue, and modifying the initialization solves that.

STATS_CHANGED