Fix lossless coding operation + cmd-line options
Lossless coding can now be triggered in two ways:
(1) Add --lossless in the command-line which will ignore any qp
parameter (preferred)
(2) Use --min-qp=<lossless_q> --max-qp<lossless_q> in the command-line
where lossless_q = 0 for 8-bit, -48 for 10-bit and -96 for 12 bit.
In this case, the encoder throws a warning prompt since the qps
are close to each other, but if you continue through the prompt
coding will work normally. The lossless check cannot be used at
this stage since that information is not available yet.

Lossless operation was broken in several ways before this patch.