Unify the quant scale for lossless to 8x unitary
Improve lossless coding performance (by up to 25%).

Previously the qindex=0 quantizer value was set as 32 - assuming the
lossless transform (WHT) has a scale of 4, combined with precision
increase by 3 bits for the quantizer. This had created a lot of
confusion in the past since qindex > 0 values assumed a transform of 8
times unitary such that 64 corresponded to unit step quantizer.
It was previously discussed in the CWG to unify lossy and lossless
transform scales to avoid this anomaly in the quantizer tables.
This table makes the desired change so that all quantization table
values assume a transform scale of 8 times unitary.

Interestingly this change seems to improve the compression efficiency
of lossless coding by about 25%. It is not exactly clear where
the improvmeent is coming from, but it is likely that identity
transforms that are also enabled for lossless coding before
were previously using mismatched scales (8 in the transform vs 4
in quantizer). This patch fixes the issue and provides a good
amount of coding gain.