Fixing QoR non-determinism caused by a data race
The issue is documented in https://gitlab.com/AOMediaCodec/avm/-/issues/501#note_2176242304. The problem was caused by an incomplete initialization of 'mbmi' member in the context of SMS which caused intra_bc to randomly be on depending on the wider context of execution. This caused non-deterministic behavior that was especially evident in the multithreaded environment as SMS is not designed/intended to work with INTRABC.