Search rectangular partitions when no valid partition can be found with --enable-rect-partitions=0
When --enable-rect-partition=0 and the frame width or height is not a multiple of the superblock size, the encoder fails with the following message.
```
Stream 0: Failed to encode frame: Unspecified internal error
    The same superblock is recoded twice. Infinite loop detected?
```
The root cause is that without rectangular partitions, a legal partitioning for the right SB column or bottom SB row cannot be found.

The fix enables rectangular partition for a SB when a legal partition cannot be found.

Fixes #498