Fix error flags at the decoder for CCSO
This is a fix for issue #543 submitted by Deepa K G.

In the encoder, when the reference frame has
different values for mi_cols or mi_rows, block
level CCSO syntax reuse is disabled, but the
frame level CCSO syntax reuse is still enabled.
However, in the decoder, error was flagged,
when frame level CCSO syntax reuse is true
and reference frame has different values for
mi_cols or mi_rows.

Also, error was not flagged at the decoder
when block level CCSO syntax reuse is true
and the reference frame has CCSO disabled.

This patch corrects the above issues.