Fix issue #556 related to nightly test failures related to CWG-E188
Some unit tests related to error resilience drops frames at the decoder to check error resiliency of the codec.

CWG-E188 uses the average of primary and secondary reference contexts to initialize the contexts of current frame.
But, if some frames are dropped at the decoder, they will not be available as a secondary reference frame at the decoder.
The encoder may still use any of those frames as secondary reference. This can cause parsing error.

This patch fixes the issue by disabling the averaging operation if an error resilient frame is seen.