Fix and clean up code comments on context derivation for signaling transform zero block flag
In issue #322, a misalignment was identified between the specification and the AVM reference software regarding the context derivation for signaling the transform zero block flag. The specification, built on top of the AV1 spec, originally used a MAX operation, while the software employed a Bitwise OR operation as an optimization. However, following the integration of CWG-D086, the MAX operation no longer equates to the Bitwise OR operation, rendering some related code comments invalid.

To resolve this, the specification has been updated to use Bitwise OR for alignment with the software. See the update here: https://gitlab.com/AOMediaCodec/av2-spec/-/commit/2c267e2b7c2a932f7c632851eef5f679fb8b9404. In this MR, outdated comments have been cleaned up, relocated as needed, and to be removed along with the macro related to CWG-D086 in future.

Closes #322