CWG-F050: Reduce Line Buffer Size from DRL and WRL
To reduce the size of the line buffer around the super-block boundary, access to information is allowed for each 8x8 unit instead of the original 4x4. The information from the bottom-left 4x4 block is used to represent the entire 8x8 block.

To achieve the goal of reducing the line buffer size, special handling is required for MV access from warp causal:
1) MV access for warp mode need to be handled the same way as in DRL.

With this proposed solution, it can reduce the DRL and the WRL line buffer size by 47.5%.
-4K: Saving of   9,792 bytes
-8K: Saving of 19,584 bytes

STATS_CHANGED