Fix bug disabling PC_WIENER in search_pc_wiener_visitor.
After the sanitizer fix in !1433, the skip_search variable is incorrectly set to false in search_pc_wiener_visitor when rtype = RESTORE_PC_WIENER, always disabling the mode. This change fixes the issue.

With the fix it became apparent that !1509 only updated parts of the code. Hence added changes to fully propagate !1509.