CWG-E066: Interleaved parsing at 64x64 level with SDP in key frames
This MR is one part of candidate proposal CWG-E066.

With SDP, the parsing of luma and chroma component is interleaved at super block level (128x128) in key frames. With this MR, the parsing of luma and chroma component is interleaved at 64x64 level in key frame.

In order to implement the luma and chroma parsing at 64x64 level, additional encoder only changes are needed for palette mode in key frames. Specifically, chroma palette information is now stored and updated during RDO and writing with a separate pointer to support luma and chroma interleaving at 64x64.

The performance of this MR was tested for all intra configuration with respect V9 anchor. It was tested on A3, A4, A5 and B2 classes. 
The results are as follows:

| Class  |    Y    |    U    |    V    | YUV     |
| ------ | ------  | ------  | ------  | ------  |
|   A3   |-0.0005% |-0.0006% |-0.0005% |-0.0005% |
|   A4   |-0.0002% |-0.0004% |-0.0004% |-0.0003% |
|   A5   |-0.0018% |-0.0030% |-0.0028% |-0.0019% |
|   B2   | 0.0028% | 0.0029% | 0.0030% | 0.0029% |