Fix single thread and multi-thread decode mismatch with SDP
Fix issue #613.

In current main branch, if SDP is used in inter frame, the stream decodes differently with single and multi threaded decoding. In multi threaded decode, there is missing code for extended SDP to switch from chroma_part to luma_part when parse_decode_flag equal to 2.

The fix in this MR is add the missing code for parse_decode_flag equal to 2 case.