Bugfix to missing picture boundary handling in one loop restoration related function
As noted in issue #684, the loop restoration function `av1_whole_frame_rect` is missing code changes required to support frame widths and heights that are not aligned to multiples of 8 (as proposed in CWG-F054). This merge request addresses the issue by adding the necessary picture boundary handling within the function, as suggested in issue #684.

STATS_CHANGED