Fix the single thread and mulit-thread mismatch issue for MORPH
In current AVM, the value of mbmi->morph_pred is reset to zero when current tree type is chroma_part, which is incorrect. Because this value is only used for luma and should not be changed for chroma_part tree.

This issue is fixed by adding the condition check of whether this tree type is chroma part or not. In addition, one redundant resetting is removed in this function.

Fixes issue #612.