Fix issue #680 - skip mode parsing dependency
This MR is used to address the issue #680.

It was found there are still parsing dependency in the functions `add_ref_mv_candidate_ctx` and `check_pos_and_get_base_pos`. This MR address the issue by adding the check skip_mode to avoid the parsing dependency. 

Fixes #680 

STATS_CHANGED